Made dtos inmutable, static factory method for creating the services
