+++ src/main/java/cool/tomi/mercadolibre/sdk/models/Authentication.java
@@ -14,24 +14,12 @@ public class Authentication {
         return accessToken;
     }
 
-    public void setAccessToken(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
     public String getRefreshToken() {
         return refreshToken;
     }
 
-    public void setRefreshToken(String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
-
     public Long getExpiresIn() {
         return expiresIn;
-    }
-
-    public void setExpiresIn(Long expiresIn) {
-        this.expiresIn = expiresIn;
     }
 
     @Override
