+++ src/test/java/cool/tomi/AppTest.java
@@ -15,7 +15,5 @@ public class AppTest extends TestCase {
         return new TestSuite(AppTest.class);
     }
 
-    public void testApp() {
-
-    }
+    public void testApp() {}
 }
